requirements.txt
streamlit>=1.28.0
openai>=1.3.0
anthropic>=0.7.0
python-dotenv>=1.0.0
streamlit-mic-recorder>=0.0.8